backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
psycopg==3.1.12
minio==7.2.0
python-dotenv==1.0.0
bcrypt==4.0.1
gunicorn==21.2.0
flasgger==0.9.7.1
sendgrid==6.10.0